Problem: CLASS says nothing about style and format of documentation sources
The original CLASS document only says that there should be a doc/
directory and several standard GNU-style text files, and that program
sources may include bits of documentation. With our recent addtions,
it also explicitly requires asciidoc as the default tool to parse text
files into something else (previously it was implicit, hardcoded into
zproject recipes). CLASS does not include any recommendations or rules
regarding documentation sources style, format and tools - and perhaps
it should not, since C code (and scaffolding) and human-readable text
are different creations.

The page with overall project contribution and cooperation rules should
at least highlight that *we* here in our 42ITy project have some standards
and tools for documentation, so it does not visibly look like patchwork
from unrelated authors that must be processed with dozens of toolkits
rather than one (or two) to release an end-user format like a web-site
or book.

Otherwise these unknown and unannounced rules/recommendations/guidelines
will be blissfully ignored, and the first thing unwitting newcomers will
read will already be a repulsive mess.

[42ity-org-website]$ grep -A1 -B1 -i doc class.asciidoc
* revised "verbatim" markup in-lined in the text, and fixed a few typos;
* requirement by default of `asciidoc` compatible markup for text files
and source comments, unless a different markup is not easily avoidable;
* added standard text-processing programs among "dependencies" for the
projects which aim to deliver rendered end-user documentation formats;
* the template README file markup made compatible with `asciidoc`.

--
"SHOULD", "SHOULD NOT", "RECOMMENDED", "MAY", and "OPTIONAL" in this
document are to be interpreted as described in IETF RFC 2119 (see
"http://tools.ietf.org/html/rfc2119[Key words for use in RFCs to
--
* An `AUTHORS` file listing all contributors to the project.
* A `doc` directory containing documentation.
* The _internal header file_ (`src/myp_classes.h`).
--

The project SHALL depend at least on `asciidoc` to render end-user formats
of documentation. It MAY depend on use of `xmlto`, `docbook`, `fo`, `a2x`,
`ghostscript` and other common backends and tools for rendering of specific
final formats. The project SHALL NOT depend on tools not available freely
to anybody as required part of its documentation processing recipes (e.g.
as a component of the `make distcheck` recipe chain).
--
NOTE: A copy of the sample below is committed into the repository as
`README-CLASS-template.asciidoc`.

--

These two files SHALL be the original documentation for the class.
Specifically, the class header SHALL define the API for the class, and
--
* Suitably-marked-up comments before a function MAY be used as source
material for reference documentation.

--

* Assertions SHALL be used for their documentary value, for example to
warn the reader, "this argument SHALL NOT be null".

------

Revert "PROBLEM: Reference to doc recommendation from CLASS"

This reverts commit e55b4cdab7b0f9ccaa63632c8495842702efdcfb :
PROBLEM: Reference to doc recommendation from CLASS

CLASS document already has a well defined reference to doc/ dicrectory
and accompanying rules regarding documentation.

This is already minimal and sufficient.
No need to duplicate things in different places.

SOLUTION: Rewrite the paragraph and remove the superfluous link.